social_media_api/requirements.txt
asgiref==3.11.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Django==6.0.2
django-csp==4.0
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
packaging==26.0
pillow==12.1.1
PyJWT==2.11.0
sqlparse==0.5.5
tzdata==2025.3
Werkzeug==3.1.4
